Se agregaron logs
se hizo la creacion del modelo logs, se implemento una funcion para guardar los logs en la base de datos y esta funcion se ejecuta en los controllers que modifican datos de la base de datos, asi tambien se creo una funcion para obtener el nombre de un producto con el id del producto

+++ server/controllers/productos.controllers.js
@@ -1,6 +1,7 @@
 import Post from "../models/Post.js";
-import { uploadImage, deleteImage } from "../libs/cloudinary.js";
-import fs from "fs-extra";
+// import { uploadImage, deleteImage } from "../libs/cloudinary.js";
+// import fs from "fs-extra";
+import { guardarLog } from "./logs.controller.js";
 
 export const obtenerProductos = async (req, res) => {
   try {
@@ -12,48 +13,88 @@ export const obtenerProductos = async (req, res) => {
 };
 export const crearProducto = async (req, res) => {
   try {
-    const { Nombre, Cantidad, CantidadMinima, Unidad, Proveedor } = req.body;
-    let image;
-    if (req.files?.image) {
-      const result = await uploadImage(req.files.image.tempFilePath);
-      await fs.remove(req.files.image.tempFilePath);
-      image = { url: result.secure_url, public_id: result.public_id };
-    }
+    const { Nombre, Cantidad, CantidadMinima, Unidad, Proveedor, Usuario } =
+      req.body;
+    // let image;
+    // if (req.files?.image) {
+    //   const result = await uploadImage(req.files.image.tempFilePath);
+    //   await fs.remove(req.files.image.tempFilePath);
+    //   image = { url: result.secure_url, public_id: result.public_id };
+    // }
     const newPost = new Post({
       Nombre,
       Cantidad,
       CantidadMinima,
       Unidad,
       Proveedor,
-      image,
     });
     await newPost.save();
+    await guardarLog(
+      Usuario,
+      "Guardar nuevo",
+      `El usuario ${Usuario} guardo en la base de datos el producto: ${Nombre}`
+    );
     res.json(newPost);
   } catch (error) {
     return res.status(500).json("error");
   }
 };
 export const actualizarProductoPorId = async (req, res) => {
+  console.log("actualizando");
   try {
-    const updatePost = await Post.updateOne({ _id: req.params.id }, req.body, {
-      new: true,
-    });
+    const { Nombre, Cantidad, CantidadMinima, Unidad, Proveedor, Usuario } =
+      req.body;
+    const updatePost = await Post.updateOne(
+      { _id: req.params.id },
+      { Nombre, Cantidad, CantidadMinima, Unidad, Proveedor },
+      {
+        new: true,
+      }
+    );
+    await guardarLog(
+      Usuario,
+      "Actualizar",
+      `El usuario ${Usuario} actualizo en la base de datos el producto: ${Nombre}`
+    );
     return res.send(updatePost);
   } catch (error) {
-    return res.status(500).json("error");
+    return res.status(500).json("error al actualizar");
   }
 };
 export const eliminarProductoPorId = async (req, res) => {
   try {
-    const postRemoved = await Post.findByIdAndDelete(req.params.id);
-    if (!postRemoved) return res.sendStatus(404);
-    if (postRemoved.image.public_id) {
-      await deleteImage(postRemoved.image.public_id);
+    console.log(req.body);
+    if (!req.body.usuario) {
+      console.log("no hay ususario");
+      return;
     }
+    const { usuario } = req.body;
+    const Nombre = await obtenerProductoPorIdFunction(req.params.id);
+    if (!Nombre) {
+      console.log("No hay nombre");
+      return;
+    }
+
+    const postRemoved = await Post.findByIdAndDelete(req.params.id);
+    if (!postRemoved) {
+      return res.sendStatus(404);
+    }
+    // if (postRemoved.image.public_id) {
+    //   await deleteImage(postRemoved.image.public_id);
+    // }
+    await guardarLog(
+      usuario,
+      "Eliminar",
+      `El usuario ${usuario} elimino en la base de datos el producto: ${Nombre}`
+    );
     return res.sendStatus(204);
   } catch (error) {
     return res.status(500).json("error");
   }
+};
+const obtenerProductoPorIdFunction = async (id) => {
+  const respuesta = await Post.findById(id);
+  return respuesta.Nombre;
 };
 export const obtenerProductoPorId = async (req, res) => {
   try {
@@ -65,7 +106,6 @@ export const obtenerProductoPorId = async (req, res) => {
       if (!respuesta) {
         return res.status(404);
       }
-      console.log(respuesta);
       res.send(respuesta);
     } else {
       return res.status(500).json("error");
@@ -77,14 +117,15 @@ export const obtenerProductoPorId = async (req, res) => {
 
 export const retirarCantidadDeProductoPorId = async (req, res) => {
   try {
-    const { Cantidad } = req.body;
     if (!req.params.id) {
       return;
     }
     if (!req.body) {
       return;
     }
-    const cantidad = Cantidad;
+    const { Cantidad, Usuario } = req.body;
+    const Nombre = await obtenerProductoPorIdFunction(req.params.id);
+    const cantidad = parseInt(Cantidad);
     const objetoActualizado = await Post.updateOne(
       { _id: req.params.id },
       { $inc: { Cantidad: -cantidad } }
@@ -92,6 +133,11 @@ export const retirarCantidadDeProductoPorId = async (req, res) => {
     if (!objetoActualizado) {
       return;
     }
+    await guardarLog(
+      Usuario,
+      "Retirar",
+      `El usuario ${Usuario} Retiro de la base de datos del producto: ${Nombre}  Cantidad: ${Cantidad}`
+    );
     return res.send(objetoActualizado);
   } catch (error) {
     return res.status(500).json("error");
@@ -100,14 +146,16 @@ export const retirarCantidadDeProductoPorId = async (req, res) => {
 
 export const aumentarCantidadDeProductoPorId = async (req, res) => {
   try {
-    const { Cantidad } = req.body;
+    console.log(req.body);
     if (!req.params.id) {
       return;
     }
     if (!req.body) {
       return;
     }
-    const cantidad = Cantidad;
+    const { Cantidad, Usuario } = req.body;
+    const Nombre = await obtenerProductoPorIdFunction(req.params.id);
+    const cantidad = parseInt(Cantidad);
     const objetoActualizado = await Post.updateOne(
       { _id: req.params.id },
       { $inc: { Cantidad: +cantidad } }
@@ -115,7 +163,12 @@ export const aumentarCantidadDeProductoPorId = async (req, res) => {
     if (!objetoActualizado) {
       return;
     }
-    return res.send(objetoActualizado);
+    await guardarLog(
+      Usuario,
+      "Agregar",
+      `El usuario ${Usuario} agrego en la base de datos del producto: ${Nombre}  Cantidad: ${Cantidad}`
+    );
+    return res.send("hola");
   } catch (error) {
     return;
   }
@@ -145,10 +198,9 @@ export const verPendientes = async (req, res) => {
         $match: { menorQueCantidadMinima: true },
       },
     ]);
-    if(consultarPendientes){
+    if (consultarPendientes) {
       return res.send(consultarPendientes);
-    }
-    else return res.send(null)
+    } else return res.send(null);
   } catch (error) {
     return;
   }
@@ -166,10 +218,9 @@ export const verPendientesNotificaciones = async (req, res) => {
         $match: { menorQueCantidadMinima: true },
       },
     ]);
-    if(consultarPendientes){
+    if (consultarPendientes) {
       return consultarPendientes;
-    }
-    else return res.send(null)
+    } else return res.send(null);
   } catch (error) {
     return;
   }
